llm_factor_system/llm_factor_system/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
schedule>=1.2.0
pyyaml>=6.0
tables>=3.8.0
pyarrow>=10.0.0
scipy>=1.7.0